+++ packages/android_alarm_manager/lib/android_alarm_manager.dart
@@ -13,7 +13,7 @@ import 'package:flutter/services.dart';
 class AndroidAlarmManager {
   static const String _channelName = 'plugins.flutter.io/android_alarm_manager';
   static const MethodChannel _channel =
-      const MethodChannel(_channelName, const JSONMethodCodec());
+      MethodChannel(_channelName, JSONMethodCodec());
 
   /// Schedules a one-shot timer to run `callback` after time `delay`.
   ///
@@ -41,8 +41,8 @@ class AndroidAlarmManager {
     Duration delay,
     int id,
     dynamic Function() callback, {
-    bool exact: false,
-    bool wakeup: false,
+    bool exact = false,
+    bool wakeup = false,
   }) async {
     final int now = new DateTime.now().millisecondsSinceEpoch;
     final int first = now + delay.inMilliseconds;
@@ -81,8 +81,8 @@ class AndroidAlarmManager {
     Duration duration,
     int id,
     dynamic Function() callback, {
-    bool exact: false,
-    bool wakeup: false,
+    bool exact = false,
+    bool wakeup = false,
   }) async {
     final int now = new DateTime.now().millisecondsSinceEpoch;
     final int period = duration.inMilliseconds;