+++ packages/android_intent/example/lib/main.dart
@@ -32,8 +32,8 @@ class MyHomePage extends StatelessWidget {
   void _createAlarm() {
     final AndroidIntent intent = const AndroidIntent(
       action: 'android.intent.action.SET_ALARM',
-      arguments: const <String, dynamic>{
-        'android.intent.extra.alarm.DAYS': const <int>[2, 3, 4, 5, 6],
+      arguments: <String, dynamic>{
+        'android.intent.extra.alarm.DAYS': <int>[2, 3, 4, 5, 6],
         'android.intent.extra.alarm.HOUR': 21,
         'android.intent.extra.alarm.MINUTES': 30,
         'android.intent.extra.alarm.SKIP_UI': true,
@@ -92,7 +92,7 @@ class ExplicitIntentsWidget extends StatelessWidget {
     intent.launch();
   }
 
-  void _displayMapInGoogleMaps({int zoomLevel: 12}) {
+  void _displayMapInGoogleMaps({int zoomLevel = 12}) {
     final AndroidIntent intent = new AndroidIntent(
         action: 'action_view',
         data: Uri.encodeFull('geo:37.7749,-122.4194?z=$zoomLevel'),