+++ packages/firebase_auth/test/firebase_auth_test.dart
@@ -20,6 +20,7 @@ const String kMockCustomToken = '12345';
 const String kMockPhoneNumber = '5555555555';
 const String kMockVerificationId = '12345';
 const String kMockSmsCode = '123456';
+const String kMockLanguage = 'en';
 
 void main() {
   group('$FirebaseAuth', () {
@@ -41,6 +42,9 @@ void main() {
             break;
           case "sendPasswordResetEmail":
           case "updateProfile":
+            return null;
+            break;
+          case "updateEmail":
             return null;
             break;
           case "fetchProvidersForEmail":
@@ -319,6 +323,25 @@ void main() {
       );
     });
 
+    test('delete', () async {
+      final FirebaseUser user = await auth.currentUser();
+      await user.delete();
+
+      expect(
+        log,
+        <Matcher>[
+          isMethodCall(
+            'currentUser',
+            arguments: null,
+          ),
+          isMethodCall(
+            'delete',
+            arguments: null,
+          ),
+        ],
+      );
+    });
+
     test('sendPasswordResetEmail', () async {
       await auth.sendPasswordResetEmail(
         email: kMockEmail,
@@ -351,6 +374,20 @@ void main() {
           },
         ),
       ]);
+    });
+
+    test('updateEmail', () async {
+      final String updatedEmail = 'atestemail@gmail.com';
+      auth.updateEmail(email: updatedEmail);
+      expect(
+        log,
+        <Matcher>[
+          isMethodCall(
+            'updateEmail',
+            arguments: <String, String>{'email': updatedEmail},
+          ),
+        ],
+      );
     });
 
     test('signInWithCustomToken', () async {
@@ -416,15 +453,31 @@ void main() {
         ],
       );
     });
+
+    test('setLanguageCode', () async {
+      await auth.setLanguageCode(kMockLanguage);
+
+      expect(
+        log,
+        <Matcher>[
+          isMethodCall(
+            'setLanguageCode',
+            arguments: <String, String>{
+              'language': kMockLanguage,
+            },
+          ),
+        ],
+      );
+    });
   });
 }
 
 Map<String, dynamic> mockFirebaseUser(
-        {String providerId: kMockProviderId,
-        String uid: kMockUid,
-        String displayName: kMockDisplayName,
-        String photoUrl: kMockPhotoUrl,
-        String email: kMockEmail}) =>
+        {String providerId = kMockProviderId,
+        String uid = kMockUid,
+        String displayName = kMockDisplayName,
+        String photoUrl = kMockPhotoUrl,
+        String email = kMockEmail}) =>
     <String, dynamic>{
       'isAnonymous': true,
       'isEmailVerified': false,