+++ packages/firebase_ml_vision/CHANGELOG.md
@@ -1,3 +1,12 @@
+## 0.1.0
+
+* **Breaking Change** Add Barcode, Face, and Label on-device detectors.
+* Remove close method.
+
+## 0.0.2
+
+* Bump Android and Firebase dependency versions.
+
 ## 0.0.1
 
 * Initial release with text detector.